+++ R/as_lt.R
@@ -47,13 +47,15 @@ lt::lt
 #'   a = safety_design$lower$bound,
 #'   b = safety_design$upper$bound
 #' )
-#' safety_power |>
-#'   as_table() |>
-#'   lt(
-#'     theta_label = I("Underlying<br>AE rate"),
-#'     prob_decimals = 3,
-#'     bound_label = c("low rate", "high rate")
-#'   )
+#' if (interactive()) {
+#'   safety_power |>
+#'     as_table() |>
+#'     lt(
+#'       theta_label = I("Underlying<br>AE rate"),
+#'       prob_decimals = 3,
+#'       bound_label = c("low rate", "high rate")
+#'     )
+#' }
 #'
 #' @exportS3Method lt::lt
 lt.gsBinomialExactTable <- function(