+++ R/as_table.R
@@ -28,9 +28,11 @@
 #'   k = length(b$n.I), theta = seq(.1, .45, .05), n.I = b$n.I,
 #'   a = b$lower$bound, b = b$upper$bound
 #' )
-#' b_power |>
-#'   as_table() |>
-#'   lt()
+#' if (interactive()) {
+#'   b_power |>
+#'     as_table() |>
+#'     lt()
+#' }
 as_table <- function(x, ...) UseMethod("as_table")
 
 #' @rdname as_table